requirements.txt
Pillow
pixsaw
cairosvg
opencv-python
opencv-contrib-python